Rename submodule ktool.ktool to ktool.kprint
