+++ pyk/src/pyk/ktool/__init__.py
@@ -1,3 +1,3 @@
 from .kompile import KompileBackend, kompile
+from .kprint import KPrint
 from .kprove import KProve
-from .ktool import KPrint
